Utility library builds on x64

+++ CamelotUtility/Include/CmBox.h
@@ -10,7 +10,7 @@ namespace CamelotEngine
     */
 	struct CM_UTILITY_EXPORT Box
 	{
-		size_t left, top, right, bottom, front, back;
+		UINT32 left, top, right, bottom, front, back;
 		/// Parameterless constructor for setting the members manually
 		Box()
 			: left(0), top(0), right(1), bottom(1), front(0), back(1)
@@ -25,7 +25,7 @@ namespace CamelotEngine
 			@note Note that the left, top, and front edges are included 
  				but the right, bottom and back ones are not.
 		*/
-		Box( size_t l, size_t t, size_t r, size_t b ):
+		Box( UINT32 l, UINT32 t, UINT32 r, UINT32 b ):
 			left(l),
 			top(t),   
 			right(r),
@@ -46,7 +46,7 @@ namespace CamelotEngine
 			@note Note that the left, top, and front edges are included 
  				but the right, bottom and back ones are not.
 		*/
-		Box( size_t l, size_t t, size_t ff, size_t r, size_t b, size_t bb ):
+		Box( UINT32 l, UINT32 t, UINT32 ff, UINT32 r, UINT32 b, UINT32 bb ):
 			left(l),
 			top(t),   
 			right(r),
@@ -65,10 +65,10 @@ namespace CamelotEngine
 		}
             
 		/// Get the width of this box
-		size_t getWidth() const { return right-left; }
+		UINT32 getWidth() const { return right-left; }
 		/// Get the height of this box
-		size_t getHeight() const { return bottom-top; }
+		UINT32 getHeight() const { return bottom-top; }
 		/// Get the depth of this box
-		size_t getDepth() const { return back-front; }
+		UINT32 getDepth() const { return back-front; }
 	};
 }
